Geonames comparison tooling
V0.1 of a tool that enables us to create a baseline against which we can
measure improvements / regressions in either the Android tzs2.dat file
or the geonames data that provides a reference data set of cities and
time zones.

The "known_diffs*" entries are currently not associated with bugs but
this will be done later in the cycle.

These results were generated with external/geonames git SHA d14440d

Bug: 163800751
Test: Ran the script
Change-Id: Id07c041a243f5a1b203298b8e2e12a3b2d739bd5