A reference AOSP LocationTimeZoneProvider/Service
A reference AOSP LocationTimeZoneProvider implementation that uses
on-device data.

The code has been split to be easily tested through the use of fake
OfflineLocationTimeZoneDelegate.Environment and GeoTimeZoneFinder impls.
A sample test is provided to demonstrate this but the tests are not
complete and will be extended in a later commit.

This is only the initial commit so there are several TODOs.

Bug: 152746105
Bug: 172546738
Test: atest OfflineLocationTimeZoneProviderTests
Change-Id: I5a38ede451dd2d7d8d47bfdd5077c8b7201a300d