Change the OfflineLocationTimeZoneDelegate impl
This implementation uses different location API calls for active
listening and caps the amount of active listening that can be performed.

After this change it is possible to declare that active listening
cannot exceed a maximum value in any given period. The approach should
also mean that the provider stays responsive in situations where
location listening returns results quickly.

Bug: 177678050
Bug: 152746105
Bug: 173201396
Test: atest OfflineLocationTimeZoneProviderTests
Test: build / boot / observation
Change-Id: Iae37576271b1b3c8749d5bdc21277f3daca97ee0