Make ReferenceWithHistory output easier to use
Make ReferenceWithHistory easier to use by outputting a timestamp for
each set() operation, and providing a stable "set id" for historic
entries. This means consecutive dump() calls will assign the same ID to
the same historic entry, enabling easier comparison.

This doesn't change the contract of the storage, just the dump() output.

This is a port of changes made to the system server equivalent of this
class with changes to account for missing internal classes. It also adds
tests.

Test: treehugger
Bug: 152746105
Change-Id: I3c4d0fdc3f52eea80e015c107e9d5bca67245cac