Redact privacy sensitive info from logcat
Add some support code for keeping PII out of logs. The idea is to make
it obvious that PII may be present and to do the safe thing with
toString(), even where it happens "by default" (i.e. as part of
appending objects to strings.

Apply the new classes to locationtzprovider/ and geotz_lookup/ so as to
remove privacy sensitive info from logcat while keeping it for dumpsys.

Test: build / boot / dumpsys / log inspection
Bug: 188792111
Change-Id: I85a0d5a30f92b65650e1a757716c2a19aa4147e3